requirements.txt
absl-py==2.1.0
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
astunparse==1.6.3
blinker==1.7.0
boto3==1.34.65
botocore==1.34.65
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
dnspython==2.6.1
docker==7.0.0
ecdsa==0.18.0
email_validator==2.1.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.110.0
Flask==3.0.2
flatbuffers==24.3.7
fonttools==4.50.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.28.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
grpcio==1.62.1
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
importlib_metadata==7.0.2
importlib_resources==6.3.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jose==1.0.0
kafka-python==2.0.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
Mako==1.3.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.3
mlflow==2.11.1
numpy==1.23.2
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.9.15
packaging==23.2
pandas==2.2.1
pillow==10.2.0
protobuf==3.19.6
psutil==5.9.8
psycopg2-binary==2.9.9
pyarrow==15.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycrypto==2.6.1
pycryptodome==3.20.0
pydantic==2.6.3
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
querystring-parser==1.2.4
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
s3transfer==0.10.1
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.20
sqlparse==0.4.4
starlette==0.36.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-macos==2.10.0
tensorflow-metal==0.6.0
termcolor==2.4.0
threadpoolctl==3.3.0
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.9.0
urllib3==1.26.18
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
xgboost==2.0.3
zipp==3.18.1
